+++ packages/project-access/test/file/file-access.test.ts
@@ -1,8 +1,9 @@
 import { join } from 'path';
 import type { Package } from '../../src';
-import { readFile, readJSON, fileExists } from '../../src/file';
+import { readFile, readJSON, fileExists, writeFile } from '../../src/file';
 import { create as createStorage } from 'mem-fs';
 import { create } from 'mem-fs-editor';
+import { promises } from 'fs';
 
 describe('fileAccess', () => {
     const memFs = create(createStorage());
@@ -53,6 +54,26 @@ describe('fileAccess', () => {
         });
     });
 
+    describe('writeFile', () => {
+        const filePath = join(__dirname, 'write-to-a-file.txt');
+        beforeEach(() => {
+            jest.resetAllMocks();
+            memFs.write(filePath, '');
+        });
+        test('Write to a file - mem-fs-editor', async () => {
+            const content = 'test-data';
+            await writeFile(filePath, content, memFs);
+            const result = memFs.read(filePath);
+            expect(result).toContain(content);
+        });
+        test('Write to a file', async () => {
+            const content = 'test-data';
+            const writeFileSpy = jest.spyOn(promises, 'writeFile').mockResolvedValue();
+            await writeFile(filePath, content);
+            expect(writeFileSpy).toHaveBeenNthCalledWith(1, filePath, content, { encoding: 'utf8' });
+        });
+    });
+
     describe('fileExists', () => {
         test('Check existing file, should return true', async () => {
             const exists = await fileExists(join(__dirname, 'file-access.test.ts'));
@@ -64,7 +85,7 @@ describe('fileAccess', () => {
             expect(exists).toBe(false);
         });
 
-        test('Check existing file in memf-fs, should return true', async () => {
+        test('Check existing file in mem-fs, should return true', async () => {
             const exists = await fileExists(memFilePath, memFs);
             expect(exists).toBe(true);
         });