+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.2.2
+
+### Patch Changes
+
+-   Updated dependencies [eb0b7b37]
+    -   @sap-ux/cds-annotation-parser@0.1.8
+    -   @sap-ux/odata-annotation-core@0.1.8
+    -   @sap-ux/odata-vocabularies@0.3.4
+
 ## 0.2.1
 
 ### Patch Changes